# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Tether-ERP Procurement API
  description: |
    Comprehensive REST API for procurement management system with approval workflows,
    budget constraints, document linking, and multi-stage approval chains.
  version: 1.0.0
  contact:
    name: Tether-ERP Support
    email: support@tether-erp.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development Server
  - url: https://api.tether-erp.com/api/v1
    description: Production Server
...
paths:
...
  /organization/departments/{departmentId}/users:
    get:
      tags:
        - Departments
      summary: Get department users
      operationId: getDepartmentUsers
      security:
        - BearerAuth: []
      parameters:
        - name: departmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of users in department
        "404":
          $ref: "#/components/responses/NotFound"
  /users/{userId}/department/{departmentId}:
    post:
      tags:
        - Departments
      summary: Assign user to department
      operationId: assignUserToDepartment
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: departmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: User assigned to department
        "404":
          $ref: "#/components/responses/NotFound"
  /users/{userId}/department:
    get:
      tags:
        - Departments
      summary: Get user's department
      operationId: getUserDepartment
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User's department
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags:
        - Departments
      summary: Remove user from department
      operationId: removeUserFromDepartment
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User removed from department
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: Resource not found